Increase the grf filepos size

+++ src/common/grfio.c
@@ -44,7 +44,7 @@ struct grf_filelist {
 	int srclen;         ///< compressed size
 	int srclen_aligned;
 	int declen;         ///< original size
-	int srcpos;         ///< position of entry in grf
+	int64 srcpos;         ///< position of entry in grf
 	int next;           ///< index of next filelist entry with same hash (-1: end of entry chain)
 	char type;
 	char fn[256-4*5];   ///< file name
